+++ tests/test_dirname_file.c
@@ -6,9 +6,9 @@
 
 void test_impl()
 {
-  char *path = fsio_basename("/");
+  char *path = fsio_dirname("/dir1/dir2/file.txt");
 
-  assert_true(!strcmp(path, ""));
+  assert_string_equal(path, "/dir1/dir2/");
 
   free(path);
 }